backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Redis & Caching
redis==5.0.1
aioredis==2.0.1

# Task Queue
celery==5.3.4
flower==2.0.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# AI Services
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2

# Data Processing
pandas==2.2.0
numpy==1.26.0
pydantic==2.5.0

# Utilities
python-dotenv==1.0.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-slugify==8.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# File Processing
pillow==10.1.0
python-magic==0.4.27

# Rate Limiting
slowapi==0.1.9

# WebSockets
websockets==12.0
python-socketio==5.10.0
